refactoring the purpose processing logic
