+++ lib/consent-processor.js
@@ -10,16 +10,16 @@ const {
   flipConsentDecision
 } = require("./consent-decisions");
 
-const { processProvision } = require("./consent-provisions");
-
-const { PURPOSE_OF_USE_SYSTEM } = require("./consent-valuesets");
+const { processProvision, matchPurposeOfUse } = require("./consent-provisions");
 
 async function processDecision(consentsBundle, query) {
   const consentDecisionPromises = consentsBundle
     .filter(isActive)
     .filter(isValid)
     .filter(entry => matchesCategory(entry, query))
-    .filter(entry => matchesPurpose(entry, query))
+    .filter(entry =>
+      matchPurposeOfUse(_.get(entry, "resource.provision"), query)
+    )
     .map(entry => consentDecision(entry, query));
 
   const consentDecisions = await Promise.all(consentDecisionPromises);
@@ -71,28 +71,6 @@ function matchesCategory(entry, query) {
   return (
     !query.category ||
     _.intersectionWith(queriedCategories, categories, _.isEqual).length
-  );
-}
-
-function matchesPurpose(entry, query) {
-  const untrimmedPurposes = _.get(entry, "resource.provision.purpose") || [];
-  const untrimmedQueriedPurposes = query.purposeOfUse
-    ? _.castArray(query.purposeOfUse)
-    : [];
-
-  const purposes = untrimmedPurposes
-    .map(purpose => _.pick(purpose, ["system", "code"]))
-    .filter(purpose => purpose);
-
-  const queriedPurposes = untrimmedQueriedPurposes.map(purpose => ({
-    system: PURPOSE_OF_USE_SYSTEM,
-    code: purpose
-  }));
-  
-  return (
-    !query.purposeOfUse ||
-    !_.get(entry, "resource.provision.purpose") ||
-    _.intersectionWith(queriedPurposes, purposes, _.isEqual).length
   );
 }
