+++ lib/consent-provisions.js
@@ -9,16 +9,24 @@ const {
 const { CONSENT_PERMIT, CONSENT_DENY } = require("./consent-decisions");
 
 function matchPurposeOfUse(provision, query) {
-  const provisionPurposes = provision.purpose;
+  const untrimmedPurposes = _.get(provision, "purpose") || [];
+  const untrimmedQueriedPurposes = query.purposeOfUse
+    ? _.castArray(query.purposeOfUse)
+    : [];
 
+  const purposes = untrimmedPurposes
+    .map(purpose => _.pick(purpose, ["system", "code"]))
+    .filter(purpose => purpose);
+
+  const queriedPurposes = untrimmedQueriedPurposes.map(purpose => ({
+    system: PURPOSE_OF_USE_SYSTEM,
+    code: purpose
+  }));
+    
   return (
-    !provision.purpose ||
-    provisionPurposes.some(provisionPurpose =>
-      _.isEqual(provisionPurpose, {
-        system: PURPOSE_OF_USE_SYSTEM,
-        code: query.purposeOfUse
-      })
-    )
+    !query.purposeOfUse ||
+    !_.get(provision, "purpose") ||
+    _.intersectionWith(queriedPurposes, purposes, _.isEqual).length
   );
 }
 
@@ -132,5 +140,6 @@ function fhirBaseFromConsentUrl(fullUrl) {
 }
 
 module.exports = {
-  processProvision
+  processProvision,
+  matchPurposeOfUse
 };
